Inserido a configuração de retornar "Não Encontrado" quando for deletar algum dado que não exista.

+++ app/src/main/java/application/controller/AlunoController.java
@@ -1,8 +1,10 @@
 package application.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import application.model.Aluno;
 import application.repository.AlunoRepository;
@@ -47,6 +50,10 @@ public class AlunoController {
 
     @DeleteMapping("/alunos/{id}")
     public void deleteAluno(@PathVariable Long id) {
-        alunoRepo.deleteById(id);
+        if (alunoRepo.existsById(id)) {
+            alunoRepo.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
